vehicle-ai-assistant/requirements.txt
flask==2.2.3
flask-cors==3.0.10
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.9
requests==2.28.2
gunicorn==20.1.0
werkzeug==2.2.3 